Moved get_mousename from plot_feats to feature_analysis_and_visualization.utils
